refactor: split middleware into individual files
